+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   * Message queues: `enqueue` now takes an optional flag to specify dedupe behaviour when message is locked.
   * Message queues: `queue-status` now shows next & last mid.
   * Message queues: add support for delayed start (use `:auto-start <ms>` option).
+  * Message queues: log notice on worker start/stop.
 
   * Tundra: completely refactored design for robustness+simplicity - now use standard message queue for replication worker.
   * Tundra: replication workers have inherited message queue features like retries, backoffs, etc.